move more leaf nodes to oas
