+++ hack/verify-imports.sh
@@ -78,6 +78,7 @@ RC=0
 print_forbidden_imports oauth-server k8s.io/kubernetes/pkg/apis || RC=1
 print_forbidden_imports oc k8s.io/kubernetes/pkg || RC=1
 print_forbidden_imports openshift-apiserver k8s.io/kubernetes/pkg/apis || RC=1
+print_forbidden_imports openshift-controller-manager k8s.io/kubernetes/pkg/api/legacyscheme k8s.io/kubernetes/pkg/apis k8s.io/kubernetes/pkg/controller k8s.io/kubernetes/pkg/registry/core/service || RC=1
 print_forbidden_imports template-service-broker k8s.io/kubernetes/pkg/apis k8s.io/kubernetes/pkg/api k8s.io/kubernetes/pkg/kubectl k8s.io/kubernetes/pkg/controller || RC=1
 if [ ${RC} != 0 ]; then
     exit ${RC}
