+++ pkg/cmd/openshift-controller-manager/controller/ingress.go
@@ -4,8 +4,8 @@ import (
 	coreclient "k8s.io/client-go/kubernetes/typed/core/v1"
 
 	routeclient "github.com/openshift/client-go/route/clientset/versioned/typed/route/v1"
+	"github.com/openshift/openshift-controller-manager/pkg/route/ingress"
 	"github.com/openshift/origin/pkg/cmd/server/bootstrappolicy"
-	"github.com/openshift/origin/pkg/route/controller/ingress"
 )
 
 func RunIngressToRouteController(ctx *ControllerContext) (bool, error) {
