+++ pkg/project/apis/project/validation/validation.go
@@ -8,9 +8,9 @@ import (
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/kubernetes/pkg/apis/core/validation"
 
+	"github.com/openshift/openshift-apiserver/pkg/apiserver/labelselector"
 	oapi "github.com/openshift/origin/pkg/api"
 	projectapi "github.com/openshift/origin/pkg/project/apis/project"
-	"github.com/openshift/origin/pkg/util/labelselector"
 )
 
 func ValidateProjectName(name string, prefix bool) []string {
