+++ staging/src/github.com/openshift/openshift-apiserver/pkg/quota/quotautil/helper.go
@@ -1,4 +1,4 @@
-package util
+package quotautil
 
 import (
 	"strings"
